requirements.txt
streamlit>=1.32.0
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
openai-whisper>=20231117
moviepy>=1.0.3
ffmpeg-python>=0.2.0
transformers>=4.40.0
torch>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
